added a bit of styling. //TODO make path dynamic and add functionality to set the path

+++ app.py
@@ -5,15 +5,22 @@ from PIL import Image
 import random
 import string
 import os
-link = sys.argv[1]
+from termcolor import colored,cprint
+if(len(sys.argv) > 1):
+    link = sys.argv[1]
+else: link = '-h'
+
 PATH = '/home/thatguy/Downloads/sterea/'
 def show_list():
     return  os.listdir(PATH)
 
 def clear_folder():
+    if (len(dir_list)==0):
+        cprint('Directory is Empty','white','on_red')
+        return
     for file in dir_list:
         os.remove(f'{PATH}{file}')
-    print('Files removed!')
+    cprint('Files removed!','black','on_cya ')
 
 
 def get_random_string(length):
@@ -32,8 +39,18 @@ def convert_to_jpg():
     im = Image.open('temp.webp').convert('RGB')
     im.save(f'{PATH}{name}.jpg','jpeg')
 
+if(link=="-h" or len(sys.argv)==0):
+    text = colored("Webp to jpg converter","green",attrs=["reverse","blink"])
+    print(text)
+    print("========")
+    cprint("Usage: Run the script with the image URL as argument","black","on_cyan")
+    print("========")
+    cprint("Available Flags:","green","on_black")
+    cprint("-h: Help ","green","on_black")
+    cprint("-l: Prints all files on directory","green","on_black")
+    cprint("-d: Deletes all files on directory","green","on_black")
 
-if(link=="-l"):
+elif(link=="-l"):
     dir_list =  show_list()
     print('Files:',dir_list)
 elif(link=="-d"):
